% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{access_remote_version_control}
\alias{access_remote_version_control}
\title{access version remote control}
\usage{
access_remote_version_control(x, ...)
}
\description{
This generic function goes to GitHub to create a Personal Access Token which
is then stored in the credential store when pasting this token into the console
message
}
\seealso{
Other version_control: 
\code{\link{include_local_version_control}()},
\code{\link{include_remote_version_control}()},
\code{\link{sync_local_version_control}()}
}
\concept{version_control}
