% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{activate}
\alias{activate}
\title{activate data project}
\usage{
activate()
}
\value{
data project class
}
\description{
This function creates a new data project class instance in the global environment,
updating the project's version
}
\details{
Data project properties
\itemize{
\item location - read only location path to the data project
\item name - read only name of the data project
\item version - read only version of the project
\item data_project - read only data project class exists in global environment
\item config - project set up to include the use of the config file to store path within the project
\item data - project set up to include data
\item databases - project set up to include the connection to databases and use of SQL queries
\item documentation - project set up to include package documentation
\item fme - project set up to include the use of FME workspaces and custom transformers
\item magrittr_pipe - project set up to use the magrittr's pipe, \%>\% in the package
\item package_version_control uses \href{https://rstudio.github.io/renv/}{renv} lock file to manage the package versions within the project
\item quality_assurance functions used for the validation and quality assurance of datasets
\item rtools - read only check that the correct version of rtools is available on the local machine
\item spreadsheets - project set up to include the use of spreadsheets
\item tasks - project uses tasks quarto document to list tasks for the project
\item typst_documents - project set up to include typst documents
\item unit_tests - project set up to include function unit tests using the \href{https://testthat.r-lib.org/}{testthat package}
\item version_control - project set up to use Git as version control system
\item vignettes - project set up to include vignette quarto documents
}

Data project generic functions and methods
\itemize{
\item access_version_control - create and store a GitHub Personal Access token in the local credential store
\item add_data_script - includes the use of data within the project, adding a data-raw script for the dataset(s)
\item add_function_script - adds a functions script in the R folder, including suggested Roxygen comments
\item add_lookup_script - adds a lookup-raw script to the data-raw folder
\item add_quarto_vignette - adds a blank Quarto vignette to the vignettes folder
\item add_typst_document - adds a blank Typst document to the typst_docs folder
\item add_unit_test - setups the data project to use unit testing, adds a unit test file to the tests folder
\item document_all_functions_and_lookups - converts all the functions and lookup documentation into corresponding
html files for documentation
\item create_new_package - opens a new package form in RStudio to run and fill in to create a new R package
within the data project
\item document_data_project - updates the data project's documentation
\item document_home - renders the home page into corresponding markdown document for documentation
\item document_products - converts products document into corresponding html file for documentation
\item document_all_functions_and_lookups - converts all the functions and lookups markdown documents into
corresponding html files for documentation
\item document_all_quarto_articles - converts all quarto articles to corresponding html files for documentation
\item document_single_quarto_article - converts single quarto article to corresponding html file for documentation
\item document_tasks - converts tasks quarto article to corresponding html file for documentation
\item include_config - included the use of the config file within the project
\item include_databases - includes the use of connecting to databases and SQL queries within the project
\item include_documentation - includes documentation in the project
\item include_fme - includes the use of FME workspaces and custom transformers within the project
\item include_magrittr_pipe - includes the use of the magrittr pipe in the project
\item include_quality_assurance - adds functions used for the validation and quality assurance of datasets
\item include_spreadsheets - includes the use of spreadsheets within the project
\item include_version_control - creates a Git and GitHub repository for the project
\item increment_major_version - increments major level of data project version, with message to update lifecycle badge
\item increment_minor_version - increments minor level of data project version, with message to update lifecycle badge
\item open_config - opens the config yaml and .Renviron files in RStudio for editing
\item open_description - opens the DESCRIPTION document in RStudio for editing
\item open_home - opens the README Rmarkdown document in RStudio for editing
\item open_pkgdown - opens the _pkgdown document in RStudio for editing
\item open_products - opens the NEWS document in RStudio for editing
\item open_tasks - opens the tasks document in RStudio for editing
\item sync_version_control - stages and commits data project files to the git repository
}
}
