% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{add_lookup_script}
\alias{add_lookup_script}
\title{add lookup script}
\usage{
add_lookup_script(x, ...)
}
\description{
This generic function adds a lookup-raw script to the data-raw folder, adding
the data folder is this folder does not exist
}
