% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{add_quarto_document}
\alias{add_quarto_document}
\title{add quarto document}
\usage{
add_quarto_document(x, ...)
}
\arguments{
\item{document_title}{character title for document}
}
\description{
This generic function adds a blank quarto document from the data project's templates
folder into the project's quarto docs folder, creating the quarto docs folder structure
if it does not already exist
}
\details{
Add the document html document link to the pkgdown's yaml file for documentation.
}
\keyword{internal}
