% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{add_quarto_vignette}
\alias{add_quarto_vignette}
\title{add quarto vignette}
\usage{
add_quarto_vignette(x, ...)
}
\arguments{
\item{vignette_title}{character title for vignette}
}
\description{
This generic function adds a blank quarto vignette from the data project's
templates folder into the project's vignette folder, creating the vignette docs
folder structure if it does not already exist
}
\details{
Quarto vignette documents used with installed Quarto version 1.5. Add the
vignettes html document link to the pkgdown's yaml file for documentation.
}
