% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{add_typst_document}
\alias{add_typst_document}
\title{add typst document}
\usage{
add_typst_document(x, ...)
}
\arguments{
\item{document_title}{character title for document}
}
\description{
This generic function adds a blank typst document from the data project's templates
folder into the project's typst_docs folder, , creating the typst_docs folder
structure if it does not already exist
}
