% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{add_unit_test}
\alias{add_unit_test}
\title{add unit tests}
\usage{
add_unit_test(x, ...)
}
\description{
This generic function adds an unit test file into the tests folder by running
\href{https://usethis.r-lib.org/reference/use_test.html}{usethis use_test function}.
The data project to set up to include unit tests if necessary,
by running \link{usethis use_data_raw function} (https://usethis.r-lib.org/reference/use_testthat.html).
}
\details{
Tests are created in the testthat sub-folder by running \href{https://usethis.r-lib.org/reference/use_r.html}{usethis use_test function}
}
