% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example}
\alias{code_example}
\title{code example}
\usage{
code_example()
}
\description{
A code example pasted into RStudio and added to the RStudio Addins, as described
in \href{https://rstudio.github.io/rstudioaddins/}{RStudio Addins}
}
\details{
This code example function is automatically added into the RStudio's Addins list
by adding the following fields in inst/rstudio/addins.dcf file

Name: Name of code example
Description: Description of code example
Binding: Code example function name
Interactive: false
}
