% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_add_list_column}
\alias{code_example_add_list_column}
\title{code example add list column}
\usage{
code_example_add_list_column(...)
}
\description{
This code example adds a list column into a data frame
}
