% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_apply_function_list_column}
\alias{code_example_apply_function_list_column}
\title{code example apply function to list column row}
\usage{
code_example_apply_function_list_column(...)
}
\description{
This code example applies a function to each row of a list column
}
