% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_create_list_column}
\alias{code_example_create_list_column}
\title{code example create list column}
\usage{
code_example_create_list_column(...)
}
\description{
This code example creates a list column within a data frame
}
