% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_extract_column_from_list_column}
\alias{code_example_extract_column_from_list_column}
\title{code example extract column from list column}
\usage{
code_example_extract_column_from_list_column(...)
}
\description{
This code example extracts a column from a list column within a data frame
}
