% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_multiple_parameterized_reports}
\alias{code_example_multiple_parameterized_reports}
\title{code example create multiple parameterized reports}
\usage{
code_example_multiple_parameterized_reports(...)
}
\description{
This code example iterates through a data frame to create multiple parameterized reports
}
