% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_select_file}
\alias{code_example_select_file}
\title{code example select file}
\usage{
code_example_select_file(...)
}
\description{
This code example adds the code to select a file path
}
