% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{code_example_tidyup_unspecified_objects}
\alias{code_example_tidyup_unspecified_objects}
\title{code example tidies up unspecified objects}
\usage{
code_example_tidyup_unspecified_objects(...)
}
\description{
This code example tidies up unspecified objects removing them from the global environment
}
