% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{create_new_package}
\alias{create_new_package}
\title{create new package}
\usage{
create_new_package()
}
\description{
This function opens a new package form in RStudio to run and fill in to create
a new R package within the data project
}
