% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{deactivate}
\alias{deactivate}
\title{deactivate data project}
\usage{
deactivate()
}
\description{
This function updates the data project prior to copying it and removing the data
project class instance from the global environment
\itemize{
\item creating or updating a renv lock file to store data project's package dependencies versions
\item refreshing the list of package dependencies in package's description file
}
}
\details{
The data project can be unzipped and copied onto a different machine, using the
\href{https://rstudio.github.io/renv/}{renv} package to manage the data project's
packages
\itemize{
\item renv::activate()
\item renv::restore(prompt = FALSE))
}
}
