% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_all_functions_and_lookups}
\alias{document_all_functions_and_lookups}
\title{document all functions and lookups}
\usage{
document_all_functions_and_lookups(x, ...)
}
\description{
This generic function builds all the functional and lookup documentation into
corresponding html documents. These documents are moved into the reference sub-folder
of the doc folder so that they can be included in the package documentation, created by
the \href{https://pkgdown.r-lib.org/}{pkgdown package}
}
