% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_all_quarto_articles}
\alias{document_all_quarto_articles}
\title{document all quarto articles}
\usage{
document_all_quarto_articles(x, ...)
}
\description{
This generic function builds all the quarto articles in the package's vignettes folder
into corresponding html articles. These articles are moved into the articles sub-folder
of the doc folder so that it can be included in package documentation, created by the
\href{https://pkgdown.r-lib.org/}{pkgdown package}
}
