% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_data_project}
\alias{document_data_project}
\title{document data project
#'}
\usage{
document_data_project(x, ...)
}
\description{
This generic function builds the site documentation in to the doc folder of the
data project, as created by the \href{https://pkgdown.r-lib.org/}{pkgdown package}
}
