% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_home}
\alias{document_home}
\title{document home page}
\usage{
document_home(x, ...)
}
\description{
This generic function renders the README home page to be included in the package
documentation, created by the \href{https://pkgdown.r-lib.org/}{pkgdown package}
}
