% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_products}
\alias{document_products}
\title{document products}
\usage{
document_products(x, ...)
}
\description{
This generic function builds the products documentation into the corresponding
html document. This document is moved into the news sub-folder of the doc folder
so that it can be included in the package documentation, created by the
\href{https://pkgdown.r-lib.org/}{pkgdown package}
}
