% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_single_quarto_article}
\alias{document_single_quarto_article}
\title{document single quarto article}
\usage{
document_single_quarto_article(x, ...)
}
\description{
This generic function builds a single selected quarto article in the package's vignettes folder
into a corresponding html article. This article is moved into the articles sub-folder
of the doc folder so that it can be included in package documentation, created by the
\href{https://pkgdown.r-lib.org/}{pkgdown package}
}
