% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{document_tasks}
\alias{document_tasks}
\title{document tasks document}
\usage{
document_tasks(x, ...)
}
\description{
This generic function builds the tasks document in the package's vignette folder
into a corresponding html article. This article is moved into the articles sub-folder
of the doc folder so that it can be included in package documentation, created by the
\href{https://pkgdown.r-lib.org/}{pkgdown package}
}
