% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-package.R
\docType{package}
\name{dp-package}
\alias{dp}
\alias{dp-package}
\title{dp: Data project}
\description{
This package contains the S7 properties, generic functions and methods used to maintain a data project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcfrench/dp}
  \item Report bugs at \url{https://github.com/gcfrench/dp/issues}
}

}
\author{
\strong{Maintainer}: Graham French \email{graham.french@jncc.gov.uk}

}
\keyword{internal}
