% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_config}
\alias{include_config}
\title{include config}
\usage{
include_config(x, ...)
}
\description{
This generic function does the necessary setup to include config file within the
project, running \href{https://usethis.r-lib.org/reference/use_package.html}{usethis use_package function}
to add config package to the DESCRIPTION file and creating config and .Renviron files
within the project. The config file is initially set up to run in a development environment.
}
\details{
The config file can be changed to a production environment by updating the R_CONFIG_ACTIVE
setting in .Renviron file to "production"
}
