% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_databases}
\alias{include_databases}
\title{include connections to database and SQL queries}
\usage{
include_databases(x, ...)
}
\description{
This generic function does the necessary setup to include connecting to databases
and SQL queries within the project
}
\details{
Database functions are added to securely store database connections in the Windows credential store,
using the \href{https://keyring.r-lib.org/}{keyring} package, and create connection strings for SQL Server,
mySQL and PostGres databases
}
