% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_documentation}
\alias{include_documentation}
\title{include documentation}
\usage{
include_documentation(x, ...)
}
\description{
This generic function does the necessary setup to allow package documentation
by running \link{usethis use_package_doc function} (https://usethis.r-lib.org/reference/use_package_doc.html)
}
\details{
In addition the following are added
\itemize{
\item README.Rmd by running the \href{https://usethis.r-lib.org/reference/use_readme_rmd.html}{usethis use_readme_rmd function}
and \href{https://devtools.r-lib.org/reference/build_rmd.html}{devtools build_readme function}
\item NEWS.md by running the \href{https://usethis.r-lib.org/reference/use_news_md.html}{usethis use_news_md function}
\item \href{https://lifecycle.r-lib.org/}{lifecycle} image files are added to the figures sub-folder of the man folder by running the \href{https://usethis.r-lib.org/reference/use_lifecycle.html}{usethis use_lifecycle function}
}
}
