% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_fme}
\alias{include_fme}
\title{include FME workspaces and custom transformers}
\usage{
include_fme(x, ...)
}
\description{
This generic function does the necessary setup to include FME workspaces and custom transformers
within the project
}
