% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_magrittr_pipe}
\alias{include_magrittr_pipe}
\title{include magrittr pipe}
\usage{
include_magrittr_pipe(x, ...)
}
\description{
This generic function does the necessary setup to allow the magrittr pipe to
be used in the package by running \href{https://usethis.r-lib.org/reference/use_pipe.html}{usethis use_pipe function}
}
