% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_package_version_control}
\alias{include_package_version_control}
\title{include package version control using renv}
\usage{
include_package_version_control(x, ...)
}
\description{
This generic function creates a renv lock file storing package versions used within
the project
}
\keyword{internal}
