% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_quality_assurance}
\alias{include_quality_assurance}
\title{include quality assurance}
\usage{
include_quality_assurance(x, ...)
}
\description{
This generic function adds the quality assurance functions, with example penguin
dataset, used in validation and quality assurance of dataset outputs
}
