% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_quarto_documents}
\alias{include_quarto_documents}
\title{include quarto documents}
\usage{
include_quarto_documents(x, ...)
}
\description{
This generic function does the necessary setup to include quarto documents and
quarto yaml file within the project
}
\keyword{internal}
