% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_remote_version_control}
\alias{include_remote_version_control}
\title{include remote version control}
\usage{
include_remote_version_control(x, ...)
}
\description{
This generic function adds the data project to a new GitHub repository and makes
an initial commit
}
\seealso{
Other version_control: 
\code{\link{access_remote_version_control}()},
\code{\link{include_local_version_control}()},
\code{\link{sync_local_version_control}()}
}
\concept{version_control}
