% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_spreadsheets}
\alias{include_spreadsheets}
\title{include spreadsheets}
\usage{
include_spreadsheets(x, ...)
}
\description{
This generic function does the necessary setup to include spreadsheets within the project
}
