% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_tasks}
\alias{include_tasks}
\title{include tasks}
\usage{
include_tasks(x, ...)
}
\description{
This generic function copies the tasks quarto document from the data project's
templates folder in the project's vignettes folder, creating the vignettes folder structure
if it does not already exist. The current date is added as a header to the tasks document
}
\details{
Add the tasks html document link to the pkgdown's yaml file for documentation.
}
\keyword{internal}
