% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_typst_documents}
\alias{include_typst_documents}
\title{include typst documents}
\usage{
include_typst_documents(x, ...)
}
\description{
This generic function does the necessary setup to include tyspt documents within the project
}
\keyword{internal}
