% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_version_control}
\alias{include_version_control}
\title{include version control}
\usage{
include_version_control(x, ...)
}
\description{
This generic function adds the data project to a Git repository, updating the
.gitignore file with files to ignore in version control and allows an initial
commit to this repository. It creates a remote data project repository on GitHub
}
\details{
In the Terminal check Git installed on local machine
\itemize{
\item where git
\item git --version
}

In the Terminal check Git credentials are present
\itemize{
\item git config --global --list
}

Set Git credentials if not present
\itemize{
\item usethis::use_git_config(user.name = "USERNAME", user.email = "EMAIL")
}

In RStudio check Git Glocal options
\itemize{
\item version control interface for RStudio projects is ticked
\item Git executable path is set to C:/Program Files/Git/bin/git.exe
}
}
\seealso{
Other version_control: 
\code{\link{access_version_control}()},
\code{\link{sync_version_control}()}
}
\concept{version_control}
