% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{include_vignettes}
\alias{include_vignettes}
\title{include vignettes}
\usage{
include_vignettes(x, ...)
}
\description{
This generic function does the necessary setup to include vignette Rmarkdown
documents within the project
}
\details{
Vignettes are created in the vignettes sub-folder by running \href{https://usethis.r-lib.org/reference/use_vignette.html}{usethis use_vignette function}
}
\keyword{internal}
