% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{increment_major_version}
\alias{increment_major_version}
\title{increment data project's major version}
\usage{
increment_major_version(x, ...)
}
\description{
This generic function increments the data project's major version level,
with message to update lifecycle badge
}
