% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{increment_minor_version}
\alias{increment_minor_version}
\title{increment data project's minor version}
\usage{
increment_minor_version(x, ...)
}
\description{
This generic function increments the data project's minor version level, with
with message to update lifecycle badge
}
