% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{increment_version}
\alias{increment_version}
\title{increment data project's version}
\usage{
increment_version(x, ...)
}
\arguments{
\item{version_type}{character level of version to increment (patch (default), minor or major)}
}
\description{
This generic function increments the data project's version level
}
\details{
The version level may be incremented by one of three levels
\itemize{
\item patch 0.0.1 to 0.0.2
\item minor 0.0.1 to 0.1.0
\item major 0.0.1 to 1.0.0
}
}
\keyword{internal}
