% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{open_description}
\alias{open_description}
\title{open DESCRIPTION document}
\usage{
open_description(x, ...)
}
\description{
This generic function opens the DESCRIPTION document in RStudio for editing
}
