% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{open_document}
\alias{open_document}
\title{open document}
\usage{
open_document(x, ...)
}
\arguments{
\item{file_path}{character path to document within the project directory}
}
\description{
This generic function opens the specified project document in RStudio for editing
}
\keyword{internal}
