% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{open_documentation}
\alias{open_documentation}
\title{open documentation}
\usage{
open_documentation(x, ...)
}
\description{
This generic function opens the data project's documentation in a web browser
}
