% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{open_quarto}
\alias{open_quarto}
\title{open quarto document}
\usage{
open_quarto(x, ...)
}
\description{
This generic function opens the quarto yml document in RStudio for editing
}
\keyword{internal}
