% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{pkgdown_database_section}
\alias{pkgdown_database_section}
\title{pkgdown database section text}
\usage{
pkgdown_database_section()
}
\value{
vector containing database section reference text
}
\description{
Vector containing the database section reference text to be inserted into the
_pkgdown.yml file
}
\keyword{internal}
