% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{pkgdown_quality_assurance_section}
\alias{pkgdown_quality_assurance_section}
\title{pkgdown quality assurance section text}
\usage{
pkgdown_quality_assurance_section()
}
\value{
vector containing quality_assurance section reference text
}
\description{
Vector containing the quality assurance section reference text to be inserted
into the _pkgdown.yml file
}
\keyword{internal}
