% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{refresh_imports_package_list}
\alias{refresh_imports_package_list}
\title{refresh imports package list}
\usage{
refresh_imports_package_list(x, ...)
}
\description{
This generic function refreshes the package list in the Imports section of the
DESCRIPTION file by using the \href{https://thinkr-open.github.io/attachment/reference/att_amend_desc.html}{attachment att_amend_desc}
function to scan through the R scripts in the R directory, vignettes in the vignettes directory
and tests in the tests directory
}
\keyword{internal}
