% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{replace_text}
\alias{replace_text}
\title{replace text}
\usage{
replace_text(original_text, new_text, section = NA_character_, ...)
}
\arguments{
\item{original_text}{vector containing the original text, created by
\href{https://readr.tidyverse.org/reference/read_lines.html}{readr read_lines} function}

\item{new_text}{vector containing new text to append or insert}

\item{section}{character text of section name before the colon, to insert after.}
}
\value{
vector combining new vector text in original text
}
\description{
This function calls the update text function to replace text in a text file
}
\keyword{internal}
