% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{sync_local_version_control}
\alias{sync_local_version_control}
\title{sync local version control}
\usage{
sync_local_version_control(x, ...)
}
\description{
This generic function stages and commits data project files to the git repository
}
\details{
Alternatively use the Terminal to sync with version control
\itemize{
\item git -add -A
\item git commit -m "COMMIT MESSAGE"
}
}
\seealso{
Other version_control: 
\code{\link{access_remote_version_control}()},
\code{\link{include_local_version_control}()},
\code{\link{include_remote_version_control}()}
}
\concept{version_control}
