% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_project.R
\name{update_text}
\alias{update_text}
\title{update text file}
\usage{
update_text(combine_type)
}
\arguments{
\item{combine_type}{character text indicating type of combination ("append", "insert")}

\item{original_text}{vector containing the original text, created by
\href{https://readr.tidyverse.org/reference/read_lines.html}{readr read_lines} function}

\item{new_text}{vector containing new text to append or insert}

\item{section}{character text of section name before the colon, to insert after.}
}
\value{
vector combining new vector text in original text
}
\description{
This private function appends or inserts text in an original file,
immediately following a specified file section.
}
\keyword{internal}
